requirements-oe.txt
# authentication
oeauth==1.1.0
pyramid_oeauth==1.2.0